% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{aggregate}
\alias{aggregate}
\alias{[[.annmatrix}
\title{Split and Aggregate annmatrix Objects by Rows and Columns}
\usage{
\method{[[}{annmatrix}(x, i, j, ..., drop = TRUE, exact = TRUE)
}
\arguments{
\item{x}{an \code{annmatrix} object}

\item{i}{vector of length \code(nrow(x)) specifying row-wise groups}

\item{j}{vector of length \code(ncol(x)) specifying column-wise groups}

\item{...}{optional arguments where first argument specifies the function to be applied on each row-colum group and remaining arguments are passed as additional arguments for this function}

\item{drop}{if TRUE (default) the result of subsetting a single row or column is returned as a vector.}

\item{exact}{if TRUE (default) the result of subsetting a single row or column is returned as a vector.}
}
\value{
An \code{annmatrix} of lists
}
\description{
A special operator for performing aggregation operatios on rows and columns of \code{annmatrix} objects.
}
\details{
\code{`[[`}, when called with 3 arguments or more aggregates the matrix into groups by rows and columns and optionally applies a function to each resulting aggregate.
}
\examples{
coldata <- data.frame(group=c(rep("case", 20), rep("control", 20)),
                      gender=sample(c("M", "F"), 40, replace=TRUE)
                      )
rowdata <- data.frame(chr=sample(c("chr1", "chr2"), 100, replace=TRUE),
                      pos=runif(100, 0, 1000000)
                      )
X <- annmatrix(mat, rowdata, coldata)

# split X by rows and by columns
X[[X@chr, X$gender,]]

# calculate mean after splitting
X[[X@chr, X$gender, mean]]

}
\author{
Karolis KonceviÄius
}
