% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{annmat2list}
\alias{annmat2list}
\title{Transform annmatrix Object to List Format}
\usage{
annmat2list(x, ...)
}
\arguments{
\item{x}{annmatrix object}

\item{...}{other parameters passed to \code{list}}
}
\value{
a \code{list}
}
\description{
Turns the matrix and it's column and row meta-data into a \code{list}.
}
\examples{
# construct the annmatrix object
coldata <- data.frame(group=c(rep("case", 20), rep("control", 20)),
                      gender=sample(c("M", "F"), 40, replace=TRUE)
                      )
rowdata <- data.frame(chr=sample(c("chr1", "chr2"), 100, replace=TRUE),
                         pos=runif(100, 0, 1000000)
                         )
X <- annmatrix(matrix(rnorm(100*40), 100, 40), rowdata, coldata)

annmat2list(X)

}
\author{
Karolis KonceviÄius
}
