% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct.R
\name{construct}
\alias{construct}
\alias{annMatrix}
\alias{is.annMatrix}
\title{Construct annMatrix object.}
\usage{
annMatrix(mat = NULL, rowAnn = NULL, colAnn = NULL)

is.annMatrix(x)
}
\arguments{
\item{mat}{a matrix}

\item{rowAnn}{annotation data.frame for rows of the mat matrix.}

\item{colAnn}{annotation data.frame for columns of the mat matrix.}

\item{x}{an R object.}
}
\value{
annMatrix() - an annMatrix object.

  is.annMatrix() - either TRUE or FALSE.
}
\description{
Functions used to construct the annMatrix object.
}
\details{
annMatrix constructs an object of class annMatrix.

is.annMatrix checks if the object is an instance of class annMatrix.
}
\examples{
  ints <- matrix(rnorm(200), ncol=10, nrow=20)
  gmap <- data.frame(chr=sample(c("chr1", "chr2"), 20, replace=TRUE))
  key  <- data.frame(Lot=sample(10, replace=TRUE), ID=10:1)
  newAnnMat <- annMatrix(ints, gmap, key)

  is.annMatrix(ints)
  is.annMatrix(newAnnMat)

}
\author{
Karolis Koncevicius
}
