% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{$.annMatrix}
\alias{$.annMatrix}
\alias{@.annMatrix}
\title{annMatrix getters}
\usage{
\method{$}{annMatrix}(annMat, cname)

\method{@}{annMatrix}(annMat, rname)
}
\arguments{
\item{annMat}{annMatrix object.}

\item{cname}{name of existing column annotation field}

\item{rname}{name of existing row annotation field}
}
\value{
One-column data frame with selected annotation.
}
\description{
Functions used to get fields of annMatrix object.
}
\details{
Returns selected fields from row and column description annotations.
$ is used for selecting column annotations and @ for row annotations.

when a special symbol "." is used instead of a name - the whole
annotation data.frame is returned
}
\examples{
  annMatExample$group
  annMatExample@chr
  annMatExample$.
  annMatExample@.

}
\author{
Karolis Koncevicius
}
