% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupgenerics.R
\name{groupgenerics}
\alias{groupgenerics}
\alias{Ops.annmatrix}
\title{Group Generic Functions for annmatrix Class}
\usage{
\method{Ops}{annmatrix}(e1, e2)
}
\arguments{
\item{e1, e2}{annmatrix objects}
}
\description{
The functions listed here work under the hood and are almost never called by the user.
}
