% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixgenerics.r
\name{matrixgenerics}
\alias{matrixgenerics}
\alias{\%*\%.annmatrix}
\title{Matrix Generic Functions for annmatrix Class}
\usage{
\method{\%*\%}{annmatrix}(x, y)
}
\arguments{
\item{x, y}{numeric or complex matrices or vectors.}
}
\description{
Matrix cross-product operator implemented for annmatrix class
}
\details{
The resulting matrix will be the same as a product between two regular matrices.
If present annmatrix row annotations will be carried over from the first matrix \code{x}
while the annotations for rows will be carried over from the second matrix \code{y}.
}
\author{
Karolis Koncevičius
}
