% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{prcomp.annmatrix}
\alias{prcomp.annmatrix}
\title{Principal Component Analysis for annmatrix}
\usage{
\method{prcomp}{annmatrix}(
  x,
  retx = TRUE,
  center = TRUE,
  scale. = FALSE,
  tol = NULL,
  rank. = NULL,
  ...
)
}
\arguments{
\item{x}{annmatrix object.}

\item{retx}{logical indicator whether the rotated variables should be returned (defaults to TRUE).}

\item{center}{logical value indicating whether variables should be centered (defaults to TRUE).}

\item{scale.}{logical value indicating whether variables should be scaled (defaults to FALSE).}

\item{tol}{tolerance value indicating magnitude below which components will be omitted.}

\item{rank.}{number specifying the maximal rank (max number of principal components to be used).}

\item{...}{arguments passed to or from other methods.}
}
\value{
prcom object with rotation and x matrices turned into annmatrix
}
\description{
Perform principal component analysis on an annmatrix object.
}
\details{
Performs PCA on annmatrix and preserves row and column annotations in scores and loadings.
}
\examples{
# construct the annmatrix object
coldata <- data.frame(group=c(rep("case", 20), rep("control", 20)),
                      gender=sample(c("M", "F"), 40, replace=TRUE)
                      )
rowdata <- data.frame(chr=sample(c("chr1", "chr2"), 100, replace=TRUE),
                         pos=runif(100, 0, 1000000)
                         )
X <- annmatrix(matrix(rnorm(100*40), 100, 40), rowdata, coldata)

prcomp(t(X))

}
\author{
Karolis KonceviÄius
}
