% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.annMatrix}
\alias{print.annMatrix}
\title{Print annMatrix object.}
\usage{
\method{print}{annMatrix}(annMat, nrow = 10, ncol = 10)
}
\arguments{
\item{annMat}{a matrix}

\item{nrow}{number of rows to display}

\item{ncol}{number of columns to display}
}
\value{
invisibly returns annMatrix object.
}
\description{
Functions to print an annMatrix object
}
\details{
annMatrix objects are printed in a shortened form (10 rows and 10 columns by default).
In addition the function displays information about available meta-data for rows and columns.
}
\examples{
  print(annMatExample)
  print(annMatExample, nrow=2, ncol=2)

}
\author{
Karolis Koncevicius
}
