% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters.R
\name{$<-.annMatrix}
\alias{$<-.annMatrix}
\alias{@<-.annMatrix}
\title{annMatrix setters}
\usage{
\method{$}{annMatrix}(annMat, cname) <- value

\method{@}{annMatrix}(annMat, rname) <- value
}
\arguments{
\item{annMat}{annMatrix object.}

\item{cname}{name of existing column annotation field}

\item{value}{assigned value}

\item{rname}{name of existing row annotation field}
}
\value{
annMatrix object with modified fields.
}
\description{
Functions used to change the fields of annMatrix object.
}
\details{
Modifies the selected fields from row and column description annotations.
$ is used for selecting column annotations and @ for row annotations.

when a special symbol "." is used instead of a name - the whole
annotation data.frame is replaced
}
\examples{
   annMatExample@newField <- 1:nrow(annMatExample)
   annMatExample$newField <- 1:ncol(annMatExample)

}
\author{
Karolis Koncevicius
}
