% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{shortcutautocomplete}
\alias{shortcutautocomplete}
\alias{.DollarNames.annmatrix}
\title{Auto Complete Functions for annmatrix Class}
\usage{
\method{.DollarNames}{annmatrix}(x, pattern = "")
}
\arguments{
\item{x}{annmatrix object}

\item{pattern}{a regular expression used to select possible auto-completion names.}
}
\description{
Function used to select autocomplete options for dollar `$` operator.
}
\author{
Karolis Koncevičius
}
