% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{shortcutgenerics}
\alias{shortcutgenerics}
\alias{@}
\alias{@.default}
\alias{@<-}
\alias{@<-.default}
\title{New Generic Functions for annmatrix Class}
\usage{
"@"(object, name)

\method{@}{default}(object, name)

@(object, name) <- value

\method{@}{default}(object, name) <- value
}
\arguments{
\item{object}{R object}

\item{name}{a character string specifying the slot to select/replace}

\item{value}{newly assigned value to replace the slot specified by \code{name}}
}
\description{
\code{@} operator in base R is used to access the slots within S4 objects.
\code{annmatrix} makes this operator generic and uses it to select row-annotations
from the \code{annmatrix} class.
}
\details{
\code{@<-} replacement in base R is used to change the values of slots within S4 objects.
\code{annmatrix} makes it generic and the same operator is then used within \code{annmatrix}
class to change the row annotation values.
}
\author{
Karolis KonceviÄius
}
