% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{shortcuts}
\alias{shortcuts}
\alias{@}
\alias{@.default}
\alias{@.annmatrix}
\alias{@<-}
\alias{@<-.default}
\alias{@<-.annmatrix}
\alias{$.annmatrix}
\alias{$<-.annmatrix}
\alias{.DollarNames.annmatrix}
\title{Convenient Getter Functions for Objects of Class annmatrix}
\usage{
object@name

object@name

\method{@}{annmatrix}(object, name)

object@name <- value

object@name <- value

\method{@}{annmatrix}(object, name) <- value

\method{$}{annmatrix}(x, name)

\method{$}{annmatrix}(x, name) <- value

\method{.DollarNames}{annmatrix}(x, pattern = "")
}
\arguments{
\item{name}{a character string specifying the name of metadata to select or replace}

\item{value}{newly assigned value to replace the field specified by \code{name}}

\item{x, }{object R object}

\item{pattern}{a regular expression used to select possible auto-completion names.}
}
\description{
Shortcuts to access and modify the row and column annotation of the \code{annmatrix} object.
}
\details{
\code{@} operator in base R is used to access the slots within S4 objects.
\code{annmatrix} makes this operator generic and uses it to select row-annotations
from the \code{annmatrix} class.

\code{@<-} replacement in base R is used to change the values of slots within S4 objects.
\code{annmatrix} makes it generic and the same operator is then used within \code{annmatrix}
class to change the row annotation values.

\code{$} is used to select data.frame fields of the column metadata.

\code{$<-} is used to change the field of a column metadata.
}
\examples{
  coldata <- data.frame(group=c(rep("case", 20), rep("control", 20)),
                        gender=sample(c("M", "F"), 40, replace=TRUE)
                        )
  rowdata <- data.frame(chr=sample(c("chr1", "chr2"), 100, replace=TRUE),
                        pos=runif(100, 0, 1000000)
                        )
  mat <- matrix(rnorm(100*40), 100, 40)
  annMat <- annmatrix(mat, rowdata, coldata)

  annMat$group
  annMat@chr

  annMat@newField <- 1:nrow(annMat)
  annMat$newField <- 1:ncol(annMat)
  annMat$newField
  annMat$newField <- NULL
  annMat$newField

}
\author{
Karolis KonceviÄius
}
