% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.R
\name{split}
\alias{split}
\alias{rowsplit}
\alias{colsplit}
\alias{rowsplit.default}
\alias{colsplit.default}
\alias{rowsplit.data.frame}
\alias{colsplit.data.frame}
\title{Split Matrix by Rows and Columns}
\usage{
rowsplit(x, group, ...)

colsplit(x, group, ...)

\method{rowsplit}{default}(x, group, ...)

\method{colsplit}{default}(x, group, ...)

\method{rowsplit}{data.frame}(x, group, ...)

\method{colsplit}{data.frame}(x, group, ...)
}
\arguments{
\item{x}{an R object. Typically a \code{matrix} or \code{data.frame} to be divided into groups.}

\item{group}{grouping variable for rows or columns used for dividing an object into groups.
Can also be a list of groups in which case their interaction is used for grouping.}

\item{...}{further arguments passed to or from methods.}
}
\value{
transposed annmatrix object
}
\description{
Splits a \code{matrix} or \code{data.frame} into multiple groups by their rows
or columns.
}
\examples{

rowsplit(iris, iris$Species)
colsplit(iris, c(1,1,1,1,2))

rowsplit(matrix(rnorm(20), nrow=5), c(1,1,2,2,3))

}
\author{
Karolis Koncevičius
}
