% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{[.annMatrix}
\alias{[.annMatrix}
\title{Subset annMatrix object.}
\usage{
\method{[}{annMatrix}(annMat, i = TRUE, j = TRUE, ...)
}
\arguments{
\item{annMat}{annMat object}

\item{i}{subset for rows}

\item{j}{subset for columns}
}
\value{
a subset of annMatrix object with preserved meta-data information
for selected rows and columns.
}
\description{
Functions used to take subsets of annMatrix objects.
}
\examples{
  annMatExample[1:2,]
  annMatExample[,1:2]
  annMatExample[1,]
  annMatExample[1,,drop=FALSE]

}
\author{
Karolis Koncevicius
}
