% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{t.annMatrix}
\alias{t.annMatrix}
\title{Transpose annMatrix object.}
\usage{
\method{t}{annMatrix}(annMat)
}
\arguments{
\item{annMat}{annMatrix object}
}
\value{
transposed annMatrix object
}
\description{
Transpose operator for annMatrix object.
}
\details{
Transposes the matrix and switches the places of row and column meta-data.
}
\examples{
  t(annMatExample)

}
\author{
Karolis Koncevicius
}
