% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{annMat2Long}
\alias{annMat2Long}
\title{Transform annMatrix object to long format}
\usage{
annMat2Long(annMat, ...)
}
\arguments{
\item{annMat}{annMatrix object}

\item{...}{other parameters passed to data.frame()}
}
\value{
data.frame
}
\description{
Turns the matrix and it's column and row meta-data into a long-format data.frame.
}
\examples{
  annMat2Long(annMatExample)

}
\author{
Karolis Koncevicius
}
