% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind.r
\name{bind}
\alias{bind}
\alias{rbind.annmatrix}
\alias{cbind.annmatrix}
\title{Bind several annmatrix Objects Together}
\usage{
\method{rbind}{annmatrix}(...)

\method{cbind}{annmatrix}(...)
}
\arguments{
\item{...}{(generalized) vector or matrix objects}
}
\value{
a single annmatrix object resulting from binding all the inputs together
}
\description{
Implementation of \code{rbind} and \code{cbind} methods for annmatrix objects.
}
\details{
All the inputs are bound together in exact same way as if all the annmatrix objects were regular matrices.
Then, the row and column annotations of the supplied annmatrix objects are merged together.
Missing annotations are filled in using 'NA' values.

More concretely, for \code{rbind} (\code{cbind}) the column (row) annotations are merged together from all the supplied annmatrix objects.
Hence the resulting column (row) annotation data.frame will be a union of all the available column (row) annotations.
In a case when two annamtrix objects have conflicting annotations (same annotation name with different values) the first occurring instance is selected and an appropriate warning is displayed.
For row (column) annotations no conflicts are possible and they are all merged together.
When some row (column) annotation table is missing a certain annotation in a particular annmatrix it is filled with missing 'NA' values.
For non-annmatrix objects all the row (column) annotations are NA.
}
\examples{
# construct annmatrix object
x <- matrix(rnorm(20*10), 20, 10)

coldata <- data.frame(group  = rep(c("case", "control"), each = 5),
                      gender = sample(c("M", "F"), 10, replace = TRUE))

rowdata <- data.frame(chr = sample(c("chr1", "chr2"), 20, replace = TRUE),
                      pos = runif(20, 0, 1000000))

X <- annmatrix(x, rowdata, coldata)

X1 <- X[1:10,]
X2 <- X[11:20,]
all.equal(X, rbind(X1, X2))


}
\author{
Karolis KonceviÄius
}
