% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{core}
\alias{core}
\alias{annmatrix}
\alias{is.annmatrix}
\alias{as.matrix.annmatrix}
\alias{[.annmatrix}
\alias{colanns}
\alias{rowanns}
\alias{colanns<-}
\alias{rowanns<-}
\title{annmatrix Objects and Basic Functionality}
\usage{
annmatrix(x, rann, cann)

is.annmatrix(x)

\method{as.matrix}{annmatrix}(x, ...)

\method{[}{annmatrix}(x, i, j, ..., drop = TRUE)

colanns(x, names)

rowanns(x, names)

colanns(x, names) <- value

rowanns(x, names) <- value
}
\arguments{
\item{x}{an R object to be converted, used or tested}

\item{rann}{annotation \code{data.frame} for rows of the \code{annmatrix} object}

\item{cann}{annotation \code{data.frame} for columns of the \code{annmatrix} object}

\item{...}{further arguments passed to or from methods}

\item{i}{subset for rows}

\item{j}{subset for columns}

\item{drop}{if TRUE (default) the result of subsetting a single row or column is returned as a vector.}

\item{names}{a character vector of existing row/column annotation names}

\item{value}{newly assigned value to row/column annotation field}
}
\value{
\code{annmatrix} - an \code{annmatrix} object

\code{is.annmatrix} - either TRUE or FALSE

\code{as.matrix} - a \code{matrix} object

\code{X@value} - a column named "value" from the row annotation \code{data.frame}

\code{X$value} - a column named "value" from the column annotation \code{data.frame}

\code{`[`} - a subset of annmatrix and reletad meta-data annotations
}
\description{
Functions used to construct, test and use the objects of class \code{annmatrix}.
}
\details{
\code{annmatrix} constructs an object of class \code{annmatrix}.
The function expects \code{x} to be a \code{matrix}
and \code{rowanns} and \code{colanns} to be of class \code{data.frame}.
If the passed objects are of a different class the function will try to
convert them via the use of \code{as.matrix} and \code{as.data.frame}.

\code{is.annmatrix} checks if the object is an instance of \code{annmatrix}.

\code{as.matrix.} will turn an \code{annmatrix} object into plain matrix.

\code{`[.`} returns a selected subset of annmatrix object.
Row and column meta-data annotations are preserved and subsetted where needed.
In the special case when only one column or row is selected - in order to be
consistent with the \code{matrix} behavior the dimensions of matrix are
dropped and a vector is returned. Just like in the case of matrices the
additional argument \code{drop=FALSE} can be provided in order to bypass this
and return a proper matrix instead.

\code{rowanns} and \code{colanns} returns the selected field from column and
row annotation \code{data.frame} respectively. When the selected field is
not specified the whole annotation \code{data.frame} is returned.

\code{rowanns<-} and \code{colanns<-} functions can be used to replace the fields from
column and row annotation \code{data.frame} respectively. When the selected field
is not specified the whole annotation \code{data.frame} is replaced.
}
\examples{
coldata <- data.frame(group=c(rep("case", 20), rep("control", 20)),
                      gender=sample(c("M", "F"), 40, replace=TRUE)
                      )
rowdata <- data.frame(chr=sample(c("chr1", "chr2"), 100, replace=TRUE),
                      pos=runif(100, 0, 1000000)
                      )
x <- matrix(rnorm(100*40), 100, 40)
X <- annmatrix(x, rowdata, coldata)

is.annmatrix(x)
is.annmatrix(X)

X[1:2,1:2]
X[1,,drop=FALSE]

rowanns(X)
colanns(X)
rowanns(X, "chr")
colanns(X, c("group", "gender"))

rowanns(X, "newField") <- 1:nrow(X)
colanns(X, "newField") <- 1:ncol(X)
colanns(X, "newField")
colanns(X, "newField") <- NULL
colanns(X, "newField")

}
\seealso{
`$.annmatrix` `@.annmatrix`
}
\author{
Karolis KonceviÄius
}
