% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{core}
\alias{core}
\alias{annmatrix}
\alias{is.annmatrix}
\alias{[.annmatrix}
\alias{$.annmatrix}
\alias{@.annmatrix}
\alias{$<-.annmatrix}
\alias{@<-.annmatrix}
\title{annmatrix Objects and Basic Functionality}
\usage{
annmatrix(x = matrix(), rowann = NULL, colann = NULL)

is.annmatrix(x)

\method{[}{annmatrix}(x, i = TRUE, j = TRUE, ..., drop = TRUE)

\method{$}{annmatrix}(x, name)

\method{@}{annmatrix}(object, name)

\method{$}{annmatrix}(x, name) <- value

\method{@}{annmatrix}(object, name) <- value
}
\arguments{
\item{x, object}{an R object to be converted, used or tested}

\item{rowann}{annotation \code{data.frame} for rows of the annmatrix object}

\item{colann}{annotation \code{data.frame} for columns of the annmatrix object}

\item{i}{subset for rows}

\item{j}{subset for columns}

\item{...}{further arguments passed to methods}

\item{drop}{if TRUE (default) the result of subsetting a single row or column is returned as a vector.}

\item{name}{a character string name of existing row/column annotation field}

\item{value}{newly assigned value to row/column annotation field}
}
\value{
\code{annmatrix} - an \code{annmatrix} object.

\code{is.annmatrix} - either TRUE or FALSE.

\code{annMat@value} - a column named "value" from the row annotation \code{data.frame}

\code{annMat$value} - a column named "value" from the column annotation \code{data.frame}

\code{`[`.annmatrix} - a subset of  annmatrix and reletad meta-data annotations
}
\description{
Functions used to construct, test and use the objects of class \code{annmatrix}.
}
\details{
\code{annmatrix} constructs an object of class \code{annmatrix}.
The function expects \code{x} to be a \code{matrix}
and \code{rowann} and \code{colann} to be of class \code{data.frame}.
If the passed objects are of a different class the function will try to
convert them via the use of \code{as.matrix} and \code{as.data.frame}.

\code{is.annmatrix} checks if the object is an instance of \code{annmatrix}.

\code{`[.annmatrix`} returns a selected subset of annmatrix object.
Row and column meta-data annotations are preserved and subsetted where needed.
In the special case when only one column or row is selected - in order to be
consistent with the \code{matrix} behavior the dimensions of matrix are
dropped and a vector is returned. Just like in the case of matrices the
additional argument \code{drop=FALSE} can be provided in order to bypass this
and return a proper matrix instead.

\code{$} and \code{@} returns the selected field from column and row annotation
\code{data.frame} respectively. When a special symbol "." is used instead of
the field name - the whole annotation \code{data.frame} is returned.

\code{$<-} and \code{@<-} functions can be used to replace the fields from
column and row annotation \code{data.frame} respectively. When a special
symbol "." is used instead of a name - the whole annotation \code{data.frame}
is replaced.
}
\examples{
  coldata <- data.frame(group=c(rep("case", 20), rep("control", 20)),
                        gender=sample(c("M", "F"), 40, replace=TRUE)
                        )
  rowdata <- data.frame(chr=sample(c("chr1", "chr2"), 100, replace=TRUE),
                        pos=runif(100, 0, 1000000)
                        )
  mat <- matrix(rnorm(100*40), 100, 40)
  annMat <- annmatrix(mat, rowdata, coldata)

  is.annmatrix(mat)
  is.annmatrix(annMat)

  annMat[1:2,1:2]
  annMat[1,,drop=FALSE]

  annMat$group
  annMat@chr
  annMat$.
  annMat@.

  annMat@newField <- 1:nrow(annMat)
  annMat$newField <- 1:ncol(annMat)
  annMat$newField
  annMat$newField <- NULL
  annMat$newField

}
\author{
Karolis KonceviÄius
}
