% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{$.annMatrix}
\alias{$.annMatrix}
\alias{@}
\alias{@.annMatrix}
\title{annMatrix getters}
\usage{
\method{$}{annMatrix}(x, name)

"@"(object, name)

\method{@}{annMatrix}(object, name)
}
\arguments{
\item{x}{annMatrix object.}

\item{name}{name of existing row/column annotation field}

\item{object}{annMatrix object.}
}
\value{
One-column data frame with selected annotation.
}
\description{
Functions used to get fields of annMatrix object.
}
\details{
Returns selected fields from row and column description annotations.
$ is used for selecting column annotations and @ for row annotations.

when a special symbol "." is used instead of a name - the whole
annotation data.frame is returned
}
\examples{
# construct the annMatrix object
coldata <- data.frame(group=c(rep("case", 20), rep("control", 20)),
                      gender=sample(c("M", "F"), 40, replace=TRUE)
                      )
rowdata <- data.frame(chr=sample(c("chr1", "chr2"), 100, replace=TRUE),
                         pos=runif(100, 0, 1000000)
                         )
annMat <- annMatrix(matrix(rnorm(100*40), 100, 40), rowdata, coldata)

annMat$group
annMat@chr
annMat$.
annMat@.

}
\author{
Karolis Koncevicius
}
