% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupgenerics.r
\name{groupgenerics}
\alias{groupgenerics}
\alias{Ops.annmatrix}
\alias{chooseOpsMethod.annmatrix}
\title{Group Generic Functions for annmatrix Class}
\usage{
\method{Ops}{annmatrix}(e1, e2)

\method{chooseOpsMethod}{annmatrix}(x, y, mx, my, cl, reverse)
}
\arguments{
\item{e1, e2}{annmatrix objects.}

\item{x, y}{The objects being dispatched on by the group generic.}

\item{mx, my}{The methods found for objects 'x' and 'y'.}

\item{cl}{The call to the group generic.}

\item{reverse}{A logical value indicating whether 'x' and 'y' are reversed from the way they were supplied to the generic.}
}
\description{
The functions listed here work under the hood and are almost never called by the user.
}
\author{
Karolis Koncevičius
}
