% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericsnew.R
\name{newgenerics}
\alias{newgenerics}
\alias{@}
\alias{@.default}
\alias{@<-}
\alias{@<-.default}
\title{New Generics For annmAtrix Class}
\usage{
object@name

object@name

object@name <- value

object@name <- value
}
\arguments{
\item{object}{R object}

\item{name}{a character string name specifying a slot}

\item{value}{newly assigned value to replace the slot specified by \code{name}}
}
\description{
Newly implemented generic methods for annmatrix objects.
}
\details{
\code{@} operator in base R is used to access the slots within S4 objects.
annmatrix makes this operator generic and uses it to select row-annotations
from the \code{annmatrix} class.

\code{@<-} replacement in base R is used to change the values of slots within S4 objects.
annmatrix makes it generic and the same operator is then used within \code{annmatrix}
class to change the row annotation values.
}
\author{
Karolis KonceviÄius
}
