% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.annMatrix}
\alias{print.annMatrix}
\title{Print annMatrix object.}
\usage{
\method{print}{annMatrix}(x, nrow = 5, ncol = 5, ...)
}
\arguments{
\item{x}{a matrix}

\item{nrow}{number of rows to display (default is 5)}

\item{ncol}{number of columns to display (default is 5)}

\item{...}{further arguments passed to or from methods}
}
\value{
invisibly returns annMatrix object.
}
\description{
Functions to print an annMatrix object
}
\details{
annMatrix objects are printed in a shortened form (5 rows and 5 columns by default).
In addition the function displays information about available meta-data for rows and columns.
}
\examples{
# construct the annMatrix object
coldata <- data.frame(group=c(rep("case", 20), rep("control", 20)),
                      gender=sample(c("M", "F"), 40, replace=TRUE)
                      )
rowdata <- data.frame(chr=sample(c("chr1", "chr2"), 100, replace=TRUE),
                         pos=runif(100, 0, 1000000)
                         )
annMat <- annMatrix(matrix(rnorm(100*40), 100, 40), rowdata, coldata)

print(annMat)
print(annMat, nrow=2, ncol=2)

}
\author{
Karolis KonceviÄius
}
