% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters.R
\name{$<-.annMatrix}
\alias{$<-.annMatrix}
\alias{@<-}
\alias{@<-.annMatrix}
\title{annMatrix setters}
\usage{
\method{$}{annMatrix}(x, name) <- value

@(object, name) <- value

\method{@}{annMatrix}(object, name) <- value
}
\arguments{
\item{x}{annMatrix object.}

\item{name}{name of existing row/column annotation field}

\item{value}{assigned value}

\item{object}{annMatrix object.}
}
\value{
annMatrix object with modified fields.
}
\description{
Functions used to change the fields of annMatrix object.
}
\details{
Modifies the selected fields from row and column description annotations.
$ is used for selecting column annotations and @ for row annotations.

when a special symbol "." is used instead of a name - the whole
annotation data.frame is replaced
}
\examples{
# construct the annMatrix object
coldata <- data.frame(group=c(rep("case", 20), rep("control", 20)),
                      gender=sample(c("M", "F"), 40, replace=TRUE)
                      )
rowdata <- data.frame(chr=sample(c("chr1", "chr2"), 100, replace=TRUE),
                         pos=runif(100, 0, 1000000)
                         )
annMat <- annMatrix(matrix(rnorm(100*40), 100, 40), rowdata, coldata)

annMat@newField <- 1:nrow(annMat)
annMat$newField <- 1:ncol(annMat)

}
\author{
Karolis Koncevicius
}
