% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{shortcuts}
\alias{shortcuts}
\alias{@.annmatrix}
\alias{@<-.annmatrix}
\alias{$.annmatrix}
\alias{$<-.annmatrix}
\title{Convenient Functions to Access Annotations of annmatrix Objects}
\usage{
\method{@}{annmatrix}(object, name)

\method{@}{annmatrix}(object, name) <- value

\method{$}{annmatrix}(x, name)

\method{$}{annmatrix}(x, name) <- value
}
\arguments{
\item{name}{a character string specifying the name of metadata to select or replace}

\item{value}{newly assigned value to replace the field specified by \code{name}}

\item{x, object}{R object}
}
\description{
Shortcuts to access and modify the row and column annotations of the \code{annmatrix} object.
}
\details{
\code{@} selects the fields of row metadata \code{data.frame}

\code{@<-} replaces the fields of row metadata \code{data.frame}

\code{$} selects the fields of column metadata \code{data.frame}

\code{$<-} replaces the fields of column metadata \code{data.frame}
}
\examples{
coldata <- data.frame(group=c(rep("case", 20), rep("control", 20)),
                      gender=sample(c("M", "F"), 40, replace=TRUE)
                      )
rowdata <- data.frame(chr=sample(c("chr1", "chr2"), 100, replace=TRUE),
                      pos=runif(100, 0, 1000000)
                      )
mat <- matrix(rnorm(100*40), 100, 40)
annMat <- annmatrix(mat, rowdata, coldata)

annMat$group
annMat@chr

annMat@newField <- 1:nrow(annMat)
annMat$newField <- 1:ncol(annMat)
annMat$newField
annMat$newField <- NULL
annMat$newField

}
\author{
Karolis KonceviÄius
}
