% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.R
\name{split}
\alias{split}
\alias{rsplit}
\alias{csplit}
\title{Split Matrix by Rows and Columns}
\usage{
rsplit(x, group)

csplit(x, group)
}
\arguments{
\item{x}{an R object. Typically a \code{matrix} or \code{data.frame} to be divided into groups.}

\item{group}{grouping variable for rows or columns used for dividing an object into groups.
Can also be a list of groups in which case their interaction is used for grouping.}
}
\value{
transposed annmatrix object
}
\description{
Splits a \code{matrix} or \code{data.frame} into multiple groups by their rows
or columns.
}
\examples{

rsplit(iris, iris$Species)
csplit(iris, c(1,1,1,1,2))

rsplit(matrix(rnorm(20), nrow=5), c(1,1,2,2,3))

}
\author{
Karolis Koncevičius
}
