% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{t.annmatrix}
\alias{t.annmatrix}
\title{Transpose annmatrix object.}
\usage{
\method{t}{annmatrix}(x)
}
\arguments{
\item{x}{annmatrix object}
}
\value{
transposed annmatrix object
}
\description{
Transpose operator for annmatrix object.
}
\details{
Transposes the matrix and switches the places of row and column meta-data.
}
\examples{
# construct the annmatrix object
coldata <- data.frame(group=c(rep("case", 20), rep("control", 20)),
                      gender=sample(c("M", "F"), 40, replace=TRUE)
                      )
rowdata <- data.frame(chr=sample(c("chr1", "chr2"), 100, replace=TRUE),
                         pos=runif(100, 0, 1000000)
                         )
X <- annmatrix(matrix(rnorm(100*40), 100, 40), rowdata, coldata)

t(X)

}
\author{
Karolis KonceviÄius
}
