% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericstech.R
\name{techgenerics}
\alias{techgenerics}
\alias{.DollarNames.annmatrix}
\title{Technical Generic Functions for annmatrix Class}
\usage{
\method{.DollarNames}{annmatrix}(x, pattern = "")
}
\arguments{
\item{x}{annmatrix object.}

\item{pattern}{a regular expression used to select the returned auto-completion terms.}
}
\description{
Generic function that help working with the annmatrix class objects easier.
These include group generics, auto-completion helpers and similar.
}
\details{
The functions listed here work under the hood and are almost never called by
the user.
}
